Get latest version of armv7 grafana tarball
Grafana doesn't seems to give a pointer to the latest armv7 tarball.

We have to find the latest tarball name by ourself.
The deduction is kind of ugly since we assume that the last tarball name
on AWS master page is the latest one.

This may broke at some point in the futur.